+++ app/src/test/java/app/knotwork/android/domain/preset/SkillReportWriterIntegrationTest.kt
@@ -295,7 +295,7 @@ class SkillReportWriterIntegrationTest {
 
         val factory = NodeExecutorFactory(
             InputNodeExecutor(),
-            OutputNodeExecutor(llmEngine, loadModelUseCase, chatRepository),
+            OutputNodeExecutor(llmEngine, loadModelUseCase, chatRepository, mockk(relaxed = true)),
             IfConditionNodeExecutor(mockk<EvaluateIfConditionUseCase>(relaxed = true)),
             toolNodeExecutor,
             liteRtNodeExecutor,